Remove NullSdkPathException. Relocate other exceptions.
