+++ src/main/java/com/google/cloud/tools/appengine/api/deploy/AppEngineStandardStaging.java
@@ -14,7 +14,7 @@
 
 package com.google.cloud.tools.appengine.api.deploy;
 
-import com.google.cloud.tools.appengine.api.exceptions.AppEngineException;
+import com.google.cloud.tools.appengine.api.AppEngineException;
 
 /**
  * Main interface for staging Java application for App Engine standard environment before
