+++ src/main/java/com/google/cloud/tools/appengine/api/genconfig/GenConfigUtility.java
@@ -14,7 +14,7 @@
 
 package com.google.cloud.tools.appengine.api.genconfig;
 
-import com.google.cloud.tools.appengine.api.exceptions.AppEngineException;
+import com.google.cloud.tools.appengine.api.AppEngineException;
 
 /**
  * Utility for generating missing configuration files for a source directory.
