+++ src/main/java/com/google/cloud/tools/appengine/api/services/AppEngineServices.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.tools.appengine.api.services;
 
-import com.google.cloud.tools.appengine.api.exceptions.AppEngineException;
+import com.google.cloud.tools.appengine.api.AppEngineException;
 
 /**
  * This set of commands can be used to manage existing App Engine services.
