+++ src/main/java/com/google/cloud/tools/appengine/cloudsdk/CloudSdkResolver.java
@@ -18,6 +18,8 @@ package com.google.cloud.tools.appengine.cloudsdk;
 
 import java.nio.file.Path;
 
+import javax.annotation.Nonnull;
+
 /**
  * Resolve paths find the CloudSdk.
  */
@@ -28,6 +30,7 @@ public interface CloudSdkResolver {
    *
    * @return Path to Google Cloud SDK or null
    */
+  @Nonnull
   Path getCloudSdkPath();
 
   /**
