+++ src/test/java/com/google/cloud/tools/appengine/cloudsdk/CloudSdkTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.when;
 
 import com.google.cloud.tools.appengine.cloudsdk.CloudSdk.Builder;
 
-import com.google.cloud.tools.appengine.api.exceptions.AppEngineException;
+import com.google.cloud.tools.appengine.api.AppEngineException;
 import com.google.cloud.tools.appengine.cloudsdk.process.ProcessOutputLineListener;
 import org.junit.Test;
 import org.junit.runner.RunWith;
